OC1 Done!!!!!!!!!!
Now OC 1 is done. The smart car is using ultrasonic sensors. It will do PID to get close to the inner wall. Time of the run is also shown.
Users do not need to reboot the robot to restart the run. They just need to press the down button again